Added GCC_NO_HIGHLIGHT env var toggle
If there is an environment in which you do not want to add highlighting
then make sure you set the GCC_NO_HIGHLIGHT environment variable and
gcc-colour will just print out normally.

An example of such an environment is the :make command in vim where
normal terminal colour codes are not respected.