Return appropriate exit code
Previously the redo build system was deciding that the error filled
compile calls were returning succuessfully. This is because the output
from g++ was being piped into the python script which returned
successfully regardless of the gcc success rate.

We switch to a method where the gcc-colour.py script calls the gcc
command as a subprocess and monitors its output and return status. It
highlights the output and preserves the return status.